Fall speed and lock timer bugfix
Rewrite the expression that calculates the gravity because it was backwards... and also wrong. Now fall speed increases with the level as appropriate and should be using delta time. Not sure it mirrors TGM completely quite yet but it's pretty close.
Reset lock timer if block falls after it was started. Not quite enough however, since it still counts until the block has actually fallen which leads to situations where the block can freeze in mid-air if it locks just before it's supposed to fall. Will need to find a better implementation of the lock timer I think.