Lock piece bugfix
Move down the updates that happen every tick to almost the bottom, because they interfered with the lock timer (and it's where they should be anyway).
Also restore the extra else under the fall_time > gravity check that resets the lock and lock_time because it turns out that it is necessary after all.